renamed noteview etc.
