+++ src/main/java/data_access/DBLocationDataAccessObject.java
@@ -11,7 +11,7 @@ import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
-import use_case.note.DataAccessException;
+import use_case.suggest_locations.DataAccessException;
 import use_case.suggest_locations.LocationDataAccessInterface;
 
 /**
