+++ src/main/java/use_case/suggest_locations/SuggestLocationsInteractor.java
@@ -5,7 +5,6 @@ import java.util.List;
 
 import entity.Place;
 import entity.PlaceFactory;
-import use_case.note.DataAccessException;
 
 /**
  * The Suggest Locations Interactor.
